risc0_zkvm_platform
Module syscall

Modules

bigint

ecall

halt

nr

reg_abi

Structs

BigIntBlobHeader

Return

Returned registers (a0, a1) from a syscall invocation.

SyscallName

A UTF-8 NUL-terminated name of a syscall with static lifetime.

Constants

DIGEST_BYTES

DIGEST_WORDS

IO_CHUNK_WORDS

Number of words in each cycle received using the SOFTWARE ecall

MAX_BUF_BYTES

MAX_BUF_WORDS

MAX_SHA_COMPRESS_BLOCKS

Functions

sys_alloc_aligned⚠

Safety

sys_alloc_wordsDeprecated

sys_argc

Retrieves the count of arguments provided to program execution.

sys_argv⚠

Retrieves the argument with arg_index, and stores as much of it as it can it in the memory at [out_words, out_words + out_nwords).

sys_bigint⚠

Safety

sys_cycle_count

sys_exit

sys_exit() causes normal process termination.

sys_fork

sys_fork() creates a new process by duplicating the calling process. The new process is referred to as the child process. The calling process is referred to as the parent process.

sys_getenv⚠

Retrieves the value of an environment variable, and stores as much of it as it can it in the memory at [out_words, out_words + out_nwords).

sys_halt

Safety

sys_input

sys_log⚠

Safety

sys_panic⚠

Safety

sys_pause⚠

Safety

sys_pipe⚠

sys_pipe() creates a pipe, a unidirectional data channel that can be used for interprocess communication. The pointer pipefd is used to return two file descriptors referring to the ends of the pipe. pipefd[0] refers to the read end of the pipe. pipefd[1] refers to the write end of the pipe. Data written to the write end of the pipe is buffered by the host until it is read from the read end of the pipe.

sys_rand⚠

Safety

sys_read⚠

Reads the given number of bytes into the given buffer, posix-style. Returns the number of bytes actually read. On end of file, returns 0.

sys_read_words⚠

Reads up to the given number of words into the buffer [recv_buf, recv_buf + nwords). Returns the number of bytes actually read. sys_read_words is a more efficient interface than sys_read, but varies from POSIX semantics. Notably:

sys_sha_buffer⚠

Safety

sys_sha_compress⚠

Safety

sys_verify_integrity⚠

Send a ReceiptClaim digest to the host to request verification.

sys_write⚠

Safety

syscall_0⚠

Invoke a raw system call

syscall_1⚠

Invoke a raw system call

syscall_2⚠

Invoke a raw system call

syscall_3⚠

Invoke a raw system call

syscall_4⚠

Invoke a raw system call

syscall_5⚠

Invoke a raw system call